+++ code/spikes/publish.001/src/ui/Fetch.mts
@@ -1,9 +1,13 @@
 export type UrlPathString = string;
 
+import { Fetch as FetchLogic } from './logic.mjs';
+
 /**
  * Tools for fetching things.
  */
 export const Fetch = {
+  ...FetchLogic,
+
   /**
    * Dynamic imports (code-splitting).
    */
@@ -12,30 +16,5 @@ export const Fetch = {
       const { Markdown } = await import('./Markdown/Markdown');
       return Markdown;
     },
-  },
-  module: {
-    async Text() {
-      const { Text } = await import('sys.text');
-      return Text;
-    },
-  },
-
-  /**
-   * Fetch the JSON at the given URL path.
-   */
-  async json<T>(path: UrlPathString) {
-    const text = await (await fetch(path)).text();
-    return JSON.parse(text) as T;
-  },
-
-  /**
-   * Fetch the "text/markdown" from the given URL path.
-   */
-  async markdown(path: UrlPathString) {
-    const Text = await Fetch.module.Text();
-    const processor = Text.Processor.markdown();
-    const res = await fetch(path);
-    const text = await res.text();
-    return processor.toHtml(text);
   },
 };
