+++ code/spikes/publish.001/src/ui/logic.mts
@@ -1,0 +1,1 @@
+export * from '../ui.logic/index.mjs';
